add tests, add calculations, move files
