+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './pair';
 export * from './swap';
-export * from './liquidity';
 export * from './farm';
