+++ src/utils/legacyInstructions.ts
@@ -1,6 +1,7 @@
 import { PublicKey, TransactionInstruction } from '@solana/web3.js';
 import { DEFAULT_FEES } from '../constants';
 
+// UNUSED...Needed?
 export function createInitSwapInstruction(params: {
   swapAccount: PublicKey;
   authority: PublicKey;
