+++ src/utils/tokenConversions.ts
@@ -1,5 +1,6 @@
 import { SarosAMMError } from './errors';
 
+// UNUSED...Needed?
 export function fromBaseUnit(value: string | number | bigint, decimals: number = 9): string {
   const valueStr = String(value);
